Narrow IGeomeanOracle return type from uint160 to uint136
Oracle.Observation stores secondsPerLiquidityCumulativeX128 as uint136
(to fit the struct in a single 256-bit slot). IGeomeanOracle was
returning uint160 for V3 compatibility, requiring a cast loop in
JBUniswapV4Hook.observe(). Now the interface matches the storage type.

Also fixes test_initializePoolFor_revertsIfAlreadySet which used the
wrong hooks address (address(0) instead of mockOracle) for the expected
pool ID.

+++ src/libraries/JBSwapLib.sol
@@ -73,7 +73,7 @@ library JBSwapLib {
 
         // Try querying the oracle hook.
         try IGeomeanOracle(address(key.hooks)).observe(key, _makeSecondsAgos(twapWindow)) returns (
-            int56[] memory tickCumulatives, uint160[] memory secondsPerLiquidityCumulativeX128s
+            int56[] memory tickCumulatives, uint136[] memory secondsPerLiquidityCumulativeX128s
         ) {
             // Compute arithmetic mean tick from tick cumulatives.
             int56 tickCumulativesDelta = tickCumulatives[1] - tickCumulatives[0];
@@ -91,7 +91,7 @@ library JBSwapLib {
             }
 
             // Compute harmonic mean liquidity from seconds-per-liquidity cumulatives.
-            uint160 secondsPerLiquidityDelta =
+            uint136 secondsPerLiquidityDelta =
                 secondsPerLiquidityCumulativeX128s[1] - secondsPerLiquidityCumulativeX128s[0];
 
             if (secondsPerLiquidityDelta > 0) {
